Do bindless-images in a better way.
Instead of creating descriptor set layouts based on the number of
images in the image manager, use 128 images as the maximum number
of bound images and fill in dummy descriptors up to that.

This means that we can create pipelines etc. before loading models and
creating ASes which is nicer imo.